convert alarmsAutoclear to vue 3 mount
